% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDclust.R
\name{Cluster_Methylation}
\alias{Cluster_Methylation}
\title{Clustering Function}
\usage{
Cluster_Methylation(Methylation_table)
}
\arguments{
\item{input}{containing c("Sample_ID","chr","start","end", "meth") columns}
}
\description{
Clustering FNXN
}
